# product-spec/generated/qgp-api-server/src/Qgp.Api/wwwroot/openapi-original.json
{
  "openapi" : "3.1.0",
  "info" : {
    "contact" : {
      "name" : "QGP Engineering (SA/TL)"
    },
    "description" : "Public REST API cho QGP (qgp-api, .NET 8). Sinh từ SDD-QGP-001 §3, map từ PRD-QGP-001 v1.0. Base path /v1. Auth Bearer (OIDC). Verb unsafe cần header Idempotency-Key. Error code theo ISC Internal Standard — API & Error Code.\n",
    "license" : {
      "name" : "Proprietary — FPT ISC (internal)",
      "url" : "https://internal.fpt/legal"
    },
    "title" : "QGP — Quality Governance Portal API",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "description" : "Prod (nội bộ)",
    "url" : "https://qgp.internal.fpt/v1"
  }, {
    "description" : "Staging",
    "url" : "https://staging.qgp.internal.fpt/v1"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
    "/documents" : {
      "get" : {
        "operationId" : "listDocuments",
        "parameters" : [ {
          "$ref" : "#/components/parameters/Query"
        }, {
          "$ref" : "#/components/parameters/FilterType"
        }, {
          "$ref" : "#/components/parameters/FilterTag"
        }, {
          "$ref" : "#/components/parameters/FilterStatus"
        }, {
          "$ref" : "#/components/parameters/Cursor"
        }, {
          "$ref" : "#/components/parameters/Limit"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/listDocuments_200_response"
                }
              }
            },
            "description" : "Trang tài liệu"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          }
        },
        "summary" : "Danh sách / filter tài liệu (KB-F-03)",
        "tags" : [ "Documents" ]
      },
      "post" : {
        "operationId" : "createDocument",
        "parameters" : [ {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateDocumentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Document"
                }
              }
            },
            "description" : "Đã tạo (Draft)"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Trùng doc_id (BR-01)"
          },
          "422" : {
            "$ref" : "#/components/responses/ValidationError"
          }
        },
        "summary" : "Tạo tài liệu với metadata (DOC-F-01, UC-01)",
        "tags" : [ "Documents" ]
      }
    },
    "/documents/{doc_id}" : {
      "get" : {
        "operationId" : "getDocument",
        "parameters" : [ {
          "$ref" : "#/components/parameters/DocId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentDetail"
                }
              }
            },
            "description" : "Tài liệu (bản Effective hoặc cảnh báo non-Effective)"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          }
        },
        "summary" : "Lấy bản Effective + badge trạng thái (DOC-F-12, UC-07/08)",
        "tags" : [ "Documents" ]
      }
    },
    "/documents/{doc_id}/versions" : {
      "get" : {
        "operationId" : "listVersions",
        "parameters" : [ {
          "$ref" : "#/components/parameters/DocId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/DocumentVersion"
                  },
                  "type" : "array",
                  "default" : null
                }
              }
            },
            "description" : "Danh sách version"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Lịch sử phiên bản (DOC-F-06, UC-09)",
        "tags" : [ "Versions" ]
      },
      "post" : {
        "operationId" : "createVersion",
        "parameters" : [ {
          "$ref" : "#/components/parameters/DocId"
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateVersionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DocumentVersion"
                }
              }
            },
            "description" : "Version mới (Draft/UnderRevision)"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          }
        },
        "summary" : "Tạo revision minor/major (DOC-F-02, UC-05/06)",
        "tags" : [ "Versions" ]
      }
    },
    "/documents/{doc_id}/diff" : {
      "get" : {
        "operationId" : "getDiff",
        "parameters" : [ {
          "$ref" : "#/components/parameters/DocId"
        }, {
          "description" : "version nguồn (vd 2.0)",
          "explode" : true,
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "description" : "version đích (vd 2.1)",
          "explode" : true,
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getDiff_200_response"
                }
              }
            },
            "description" : "Kết quả diff"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Diff giữa 2 version (DOC-F-06)",
        "tags" : [ "Versions" ]
      }
    },
    "/versions/{id}/submit" : {
      "post" : {
        "operationId" : "submitVersion",
        "parameters" : [ {
          "$ref" : "#/components/parameters/VersionId"
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/VersionOk"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "409" : {
            "$ref" : "#/components/responses/StateConflict"
          }
        },
        "summary" : "Submit review (DOC-F-04, UC-01) — Draft → InReview",
        "tags" : [ "Versions" ]
      }
    },
    "/versions/{id}/approve" : {
      "post" : {
        "operationId" : "approveVersion",
        "parameters" : [ {
          "$ref" : "#/components/parameters/VersionId"
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/approveVersion_request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "$ref" : "#/components/responses/VersionOk"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "409" : {
            "$ref" : "#/components/responses/StateConflict"
          }
        },
        "summary" : "Approve / Reject kèm nhận xét (DOC-F-04, UC-02)",
        "tags" : [ "Versions" ]
      }
    },
    "/versions/{id}/publish" : {
      "post" : {
        "operationId" : "publishVersion",
        "parameters" : [ {
          "$ref" : "#/components/parameters/VersionId"
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "future_effective" : {
                  "summary" : "effective_date tương lai (Approved ≠ Effective)",
                  "value" : {
                    "issue_date" : "2026-07-01",
                    "effective_date" : "2026-07-15",
                    "change_summary" : "Cap nhat tieu chi Quality Gate 1"
                  }
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/PublishRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PublishResult"
                }
              }
            },
            "description" : "Đã publish"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Version đã publish — bất biến (BR-03)"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            },
            "description" : "Thiếu change_summary (BR-04) hoặc effective_date < issue_date (BR-07)"
          }
        },
        "summary" : "Publish với ngày ban hành/áp dụng (DOC-F-03, UC-03, BR-04/BR-07)",
        "tags" : [ "Versions" ]
      }
    },
    "/versions/{id}/acknowledge" : {
      "post" : {
        "operationId" : "acknowledgeVersion",
        "parameters" : [ {
          "$ref" : "#/components/parameters/VersionId"
        }, {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Acknowledgement"
                }
              }
            },
            "description" : "Đã ghi nhận acknowledgement"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Xác nhận đã đọc tài liệu bắt buộc (DOC-F-09, UC-10)",
        "tags" : [ "Versions" ]
      }
    },
    "/search" : {
      "get" : {
        "operationId" : "search",
        "parameters" : [ {
          "$ref" : "#/components/parameters/Query"
        }, {
          "$ref" : "#/components/parameters/FilterTag"
        }, {
          "$ref" : "#/components/parameters/FilterType"
        }, {
          "$ref" : "#/components/parameters/Cursor"
        }, {
          "$ref" : "#/components/parameters/Limit"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/search_200_response"
                }
              }
            },
            "description" : "Kết quả tìm kiếm (chỉ tài liệu Effective — BR-06)"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Full-text + filter, < 1s (KB-F-03, NFR-02)",
        "tags" : [ "Search" ]
      }
    },
    "/recommendations" : {
      "get" : {
        "description" : "Gợi ý explainable — chỉ tài liệu Effective trong quyền đọc (BR-11).",
        "operationId" : "getRecommendations",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Recommendations"
                }
              }
            },
            "description" : "Nhóm bắt buộc + gợi ý"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Goi y Tai lieu danh cho ban theo role (REC-F-02, UC-23)",
        "tags" : [ "Recommendations" ]
      }
    },
    "/feedback" : {
      "post" : {
        "operationId" : "createFeedback",
        "parameters" : [ {
          "$ref" : "#/components/parameters/IdempotencyKey"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateFeedbackRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Feedback"
                }
              }
            },
            "description" : "Đã tạo feedback (New)"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Gửi feedback in-context (FBK-F-01, UC-15)",
        "tags" : [ "Feedback" ]
      }
    },
    "/feedback/{id}" : {
      "patch" : {
        "operationId" : "triageFeedback",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/triageFeedback_request"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Feedback"
                }
              }
            },
            "description" : "Đã cập nhật"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          }
        },
        "summary" : "Triage feedback (FBK-F-03) — QA",
        "tags" : [ "Feedback" ]
      }
    },
    "/reports/issuance" : {
      "get" : {
        "operationId" : "getIssuanceReport",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { }
                }
              }
            },
            "description" : "Số liệu aggregate"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Thống kê tài liệu ban hành (RPT-F-01, UC-22)",
        "tags" : [ "Reports" ]
      }
    },
    "/reports/compliance" : {
      "get" : {
        "operationId" : "getComplianceReport",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "additionalProperties" : { }
                }
              }
            },
            "description" : "Tỉ lệ ack + ai chưa đọc"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Compliance acknowledgement (RPT-F-04)",
        "tags" : [ "Reports" ]
      }
    },
    "/audit" : {
      "get" : {
        "operationId" : "listAudit",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "document_id",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "from",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "to",
          "required" : false,
          "schema" : {
            "format" : "date-time",
            "type" : "string"
          },
          "style" : "form"
        }, {
          "$ref" : "#/components/parameters/Cursor"
        }, {
          "$ref" : "#/components/parameters/Limit"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/listAudit_200_response"
                }
              }
            },
            "description" : "Trang audit"
          },
          "400" : {
            "$ref" : "#/components/responses/DefaultError"
          }
        },
        "summary" : "Truy vết audit log (ADM-F-03, UC-21, NFR-03)",
        "tags" : [ "Audit" ]
      }
    },
...
  },
  "components" : {
    "parameters" : {
...
      "Query" : {
        "description" : "từ khoá tìm kiếm",
        "explode" : true,
        "in" : "query",
        "name" : "q",
        "required" : false,
        "schema" : {
          "type" : "string"
        },
        "style" : "form"
      },
      "FilterType" : {
        "explode" : true,
        "in" : "query",
        "name" : "type",
        "required" : false,
        "schema" : {
          "$ref" : "#/components/schemas/DocumentType"
        },
        "style" : "form"
      },
      "FilterTag" : {
        "explode" : true,
        "in" : "query",
        "name" : "tag",
        "required" : false,
        "schema" : {
          "type" : "string"
        },
        "style" : "form"
      },
      "FilterStatus" : {
        "explode" : true,
        "in" : "query",
        "name" : "status",
        "required" : false,
        "schema" : {
          "$ref" : "#/components/schemas/VersionStatus"
        },
        "style" : "form"
      },
      "Cursor" : {
        "description" : "opaque cursor (§3.5)",
        "explode" : true,
        "in" : "query",
        "name" : "cursor",
        "required" : false,
        "schema" : {
          "type" : "string"
        },
        "style" : "form"
      },
      "Limit" : {
        "explode" : true,
        "in" : "query",
        "name" : "limit",
        "required" : false,
        "schema" : {
          "default" : 50,
          "maximum" : 200,
          "type" : "integer"
        },
        "style" : "form"
      }
    },
    "responses" : {
      "Unauthorized" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        },
        "description" : "Thiếu/không hợp lệ token"
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        },
        "description" : "Sai role (RBAC §10.1)"
      },
...
    },
    "schemas" : {
...
      "listDocuments_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/Page"
        }, {
          "properties" : {
            "data" : {
              "items" : {
                "$ref" : "#/components/schemas/DocumentSummary"
              },
              "type" : "array",
              "default" : null
            }
          }
        } ],
        "example" : {
          "next_cursor" : "next_cursor",
          "data" : [ {
            "doc_id" : "doc_id",
            "title" : "title",
            "type" : "Policy",
            "classification" : "classification",
            "effective_version" : "effective_version",
            "status" : "Draft",
            "tags" : [ "tags", "tags" ]
          }, {
            "doc_id" : "doc_id",
            "title" : "title",
            "type" : "Policy",
            "classification" : "classification",
            "effective_version" : "effective_version",
            "status" : "Draft",
            "tags" : [ "tags", "tags" ]
          } ]
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "OIDC access token từ SSO nội bộ (ADM-F-05, NFR-04)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}